added user service and project service

+++ src/main/java/com/sit/simpleissuetracker/repository/CommentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
     public List<Comment> findByIssueIdOrderByCreatedAtAsc(Long issueId);
 }
